5.1.1
- Changed `PacketHeaderOffset` enum to use `System.UInt32`.
- Updated null argument handling in `DispatchChannel.cs`.
- Refactored `HandlePacket` methods for better resource management.
- Incremented version numbers in `Nalix.SDK.csproj` to 5.0.8.
- Modified `ReadMagicNumberLE` to use an offset for clarity.
- Enhanced `CheckSize` method for improved safety.
- Adjusted `Span` properties in `BufferLease.cs` for compatibility.
- Improved exception handling in `Retain` method of `BufferLease.cs`.
- Updated `Handshake` constructor for better clarity.
- Refactored `UnmanagedFormatter` for explicit type handling.
- Added new tests for `Handshake` serialization/deserialization.

+++ src/Nalix.Common/Packets/Enums/PacketHeaderOffset.cs
@@ -8,7 +8,7 @@ namespace Nalix.Common.Packets.Enums;
 /// Represents the positions of fields in the serialization order.
 /// Each value corresponds to a specific position in the serialized data.
 /// </summary>
-public enum PacketHeaderOffset : System.Byte
+public enum PacketHeaderOffset : System.UInt32
 {
     /// <summary>
     /// Represents the magic number field, which uniquely identifies the packet format or protocol.
